requirements.txt
minio==7.2.8
python-dotenv==1.0.1
requests==2.32.3
